content/media/webaudio/test/test_bug839753.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Crashtest for bug 839753</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

(new AudioContext()).destination.expando = null;
ok(true, "The test should not trigger wrapper cache assertions");

</script>
</pre>
</body>
</html>
